refactored user related functions to user controller and related imp files
