+++ implementation/backend/retro/src/main/java/com/cmpe451/retro/controllers/AuthenticationController.java
@@ -36,34 +36,6 @@ public class AuthenticationController {
 
     }
 
-    @RequestMapping(value = "/user", method = RequestMethod.GET)
-    public User getCurrentUser() { return authenticationService.getCurrentUser(); }
 
-    @RequestMapping(value = "/user/{id}", method = RequestMethod.GET)
-    public User getUserById(@PathVariable("id") long id) { return authenticationService.getUserById(id); }
-
-    //TODO: decide pathvariable or requestparam
-    //@RequestMapping(value = "/user/{id}", method = RequestMethod.GET)
-    //public User getUserById(@RequestParam("id") long id) { return authenticationService.getUserById(id); }
-
-    @RequestMapping(value = "/user/all", method = RequestMethod.GET)
-    public List<User> getAllUsers() { return authenticationService.getAllUsers(); }
-
-    @RequestMapping(value = "/user/info", method = RequestMethod.PUT)
-    void updateInfo(@RequestBody UpdateUserInfoRequestBody updateUserInfoBody) {
-        long userId = (long)httpServletRequest.getSession().getAttribute(Constants.USER_ID_SESSION_ATTRIBUTE);
-        authenticationService.updateUserInfo(userId, updateUserInfoBody);
-    }
-
-//    @RequestMapping(value = "/info", method = RequestMethod.GET)
-//    public UserInfoResponse getInfo(String nickname) {
-//        if (nickname == null) {
-//            long userId = sessionUtil.getUserId(httpServletRequest);
-//            return userService.getInfo(userId);
-//        }
-//
-//        return userService.getInfo(nickname);
-//
-//    }
 
 }
